Split up the notices tests
